Added Moves (Halfway) & Few Globals Vars
Added edgeDistance calculator & and the directional offsets as globals vars to be used for bishop, rook, queen, and king movement. Added the check of legality (make sure piece isn't pinned or king isn't in check). Added a Moves datasturture to help see where a piece is going to and where from. And added generate moves to generate moves for a single square. Also created function placeholders for the pieces moves to be added later.